+++ docs/conf.py
@@ -86,7 +86,7 @@ html_theme = 'flask'
 html_theme_options = {
     'github_user': 'greyli',
     'index_logo': 'flask-ckeditor.png',
-    'github_fork': 'greyli/flask-ckeditor',
+    'github_fork': 'helloflask/flask-ckeditor',
 }
 
 # Add any paths that contain custom static files (such as style sheets) here,