+++ examples/flask-admin-upload/app.py
@@ -58,21 +58,22 @@ def init_db():
     """
     Populate a small db with some example entries.
     """
-    db.drop_all()
-    db.create_all()
+    with app.app_context():
+        db.drop_all()
+        db.create_all()
 
-    # Create sample Post
-    title = "de Finibus Bonorum et Malorum - Part I"
-    text = "Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod tempor \
-                incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud \
-                exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure \
-                dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. \
-                Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt \
-                mollit anim id est laborum."
+        # Create sample Post
+        title = "de Finibus Bonorum et Malorum - Part I"
+        text = "Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod tempor \
+                    incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud \
+                    exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure \
+                    dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. \
+                    Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt \
+                    mollit anim id est laborum."
 
-    post = Post(title=title, text=text)
-    db.session.add(post)
-    db.session.commit()
+        post = Post(title=title, text=text)
+        db.session.add(post)
+        db.session.commit()
 
 
 @app.route('/files/<filename>')